Modernize the hash module
It seems nowadays libcrypto wants us to

- perform "explicit [algorithm] fetching,"
- cache algorithms to improve performance,
- and stop hardcoding algorithms in the code (which largely doesn't
  really apply in our RFC-bound situation).

https://www.openssl.org/docs/man3.0/man7/crypto.html#ALGORITHM-FETCHING

So fetch explicitely, cache algorithms and try to keep hardcoding at a
minimum.

Also, the code was seemingly unaware that hash size depends on
algorithm, so this was often validated weakly.

+++ src/slurm/db_slurm.h
@@ -30,8 +30,12 @@ struct slurm_bgpsec {
 };
 
 struct slurm_file_csum {
+	/*
+	 * Actual length is SHA256_DIGEST_LENGTH, but the documentation of
+	 * EVP_DigestFinal_ex() doesn't say we can allocate less.
+	 */
 	unsigned char csum[EVP_MAX_MD_SIZE];
-	unsigned int csum_len;
+	size_t csum_len;
 	SLIST_ENTRY(slurm_file_csum) next;
 };
